Added shortcut handling for Google Drive
-GetChildrenByParentItem now resolves (aka "passes-through") shortcuts.
-That is, it knows how to identify shortcuts and returns the object they point to rather than the zero-byte, fairly useless shortcut itself.
-Shortcuts appear as the type (file vs. folder) of the underlying target and can be navigated and chosen in the open from cloud dialog.
-Added shortcut dereferencing to GetFileByPath. This makes shortcuts work with opening and saving.
Note that copy and delete might not do what you expect. They will operate on the final target, not on the shortcut itself.
In particular, if you try to delete a shortcut to a folder, KeeAnywhere will instead try to delete the underlying real folder!
This will fail for folders not owned by you (i.e. shared with you, with a shortcut in My Drive), but it may succeed for your own shortcut-ed (i.e. symlinked) folders!
Be warned! This behavior should probably be considered a defect and fixed.