Fixed delete, so that it acts on the shortcut itself instead of the underlying file/folder.
Added a resolveFinalShortcut parameter to GetFileByPath so that we can select whether or not the last element in a path should be resolved or not.
If this parameter is set to true, shortcuts are always resolved, and the underlying file is always returned.
If this parameter is set to false, shortcuts in the path are resolved, but not the last element.